JM: started to split out to individual classes
