+++ bin/EQAlerter-Discord.py
