+++ bin/EQAlerter.py
