+++ eq_alerter/BaseMessage.py
@@ -1,0 +1,4 @@
+class BaseMessage:
+
+    def run(self):
+        subprocess.Popen(self.command, close_fds=True)
