+++ eq_alerter/CharacterList.py
